+++ src/basic/domain/cart/utils/index.ts
@@ -1,0 +1,6 @@
+// Cart domain utilities
+export {
+  applyBulkDiscount,
+  calculateItemPrice,
+  hasBulkPurchase,
+} from "./calculations";
