+++ src/basic/domain/product/utils/index.ts
@@ -1,0 +1,5 @@
+// Product domain utilities
+export {
+  calculateProductDiscount,
+  calculateRemainingStock,
+} from "./calculations";
