+++ src/basic/shared/utils/index.ts
@@ -1,0 +1,6 @@
+/**
+ * Shared utilities - 엔티티와 무관한 범용 함수들
+ */
+
+// Formatting utilities (순수 함수)
+export { formatUserPrice, formatAdminPrice, SOLD_OUT_TEXT } from "./formatters";
