+++ src/basic/utils/index.ts
@@ -1,12 +1,19 @@
-// Calculation utilities
+// Product domain utilities (migrated to domain/product/utils)
 export {
   calculateProductDiscount,
+  calculateRemainingStock,
+} from "../domain/product/utils";
+
+// Cart domain utilities (migrated to domain/cart/utils)
+export {
   applyBulkDiscount,
   calculateItemPrice,
-} from "./calculations";
+  hasBulkPurchase,
+} from "../domain/cart/utils";
 
-// Formatting utilities
-export { formatUserPrice, formatAdminPrice, SOLD_OUT_TEXT } from "./formatters";
-
-// Validation utilities
-export { hasBulkPurchase, calculateRemainingStock } from "./validators";
+// Shared utilities (migrated to shared/utils)
+export {
+  formatUserPrice,
+  formatAdminPrice,
+  SOLD_OUT_TEXT,
+} from "../shared/utils";
